Gone back to the function table but its cleaner and totally optional.
