+++ pychonet/HomeAirConditioner.py
@@ -99,131 +99,6 @@ class HomeAirConditioner(EchonetInstance):
         self.eojcc = 0x30 # Home air conditioner class
         EchonetInstance.__init__(self, self.eojgc, self.eojcc, instance, netif)
 
-    def _30A0(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-           0x41: 'auto',
-           0x31: 'minimum',
-           0x32: 'low',
-           0x33: 'medium-low',
-           0x34: 'medium',
-           0x35: 'medium-high',
-           0x36: 'high',
-           0x37: 'very-high',
-           0x38: 'max'
-        }
-        return values.get(op_mode, "Invalid setting")
-
-    # Automatic control of air flow direction setting
-    def _30A1(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-           0x41: 'auto',
-           0x42: 'non-auto',
-           0x43: 'auto-vert',
-           0x44: 'auto-horiz'
-        }
-        return values.get(op_mode, "Invalid setting")
-
-    def _30AA(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        # print(hex(op_mode))
-        values = {
-          0x40: 'Normal operation',
-          0x41: 'Defrosting',
-          0x42: 'Preheating',
-          0x43: 'Heat removal'
-          }
-        # return({'special':hex(op_mode)})
-        return values.get(op_mode, "invalid_setting")
-
-    # Automatic swing of air flow direction setting
-    def _30A3(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-           0x31: 'not-used',
-           0x41: 'vert',
-           0x42: 'horiz',
-           0x43: 'vert-horiz'
-        }
-        return values.get(op_mode, "invalid_setting")
-
-    # Air flow direction (vertical) setting
-    def _30A4(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-          0x41: 'upper',
-          0x44: 'upper-central',
-          0x43: 'central',
-          0x45: 'lower-central',
-          0x42: 'lower'
-          }
-        # return({'special':hex(op_mode)})
-        return values.get(op_mode, "invalid_setting")
-
-    # Air flow direction (horiziontal) setting
-    def _30A5(edt):
-        # complies with version 2.01 Release a (page 3-88)
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-          0x41: 'rc-right',
-          0x42: 'left-lc',
-          0x43: 'lc-center-rc',
-          0x44: 'left-lc-rc-right',
-          0x51: 'right',
-          0x52: 'rc',
-          0x54: 'center',
-          0x55: 'center-right',
-          0x56: 'center-rc',
-          0x57: 'center-rc-right',
-          0x58: 'lc',
-          0x59: 'lc-right',
-          0x5A: 'lc-rc',
-          0x60: 'left',
-          0x61: 'left-right',
-          0x62: 'left-rc',
-          0x63: 'left-rc-right',
-          0x64: 'left-center',
-          0x65: 'left-center-right',
-          0x66: 'left-center-rc',
-          0x67: 'left-center-rc-right',
-          0x69: 'left-lc-right',
-          0x6A: 'left-lc-rc'
-          }
-        # return({'special':hex(op_mode)})
-        return values.get(op_mode, "invalid_setting")
-
-    # Operation mode
-    def _30B0(edt):
-        op_mode = int.from_bytes(edt, 'big')
-        values = {
-           0x41: 'auto',
-           0x42: 'cool',
-           0x43: 'heat',
-           0x44: 'dry',
-           0x45: 'fan_only',
-           0x40: 'other'
-        }
-        return values.get(op_mode, "invalid_setting")
-
-    # Check status of Temperature
-    def _30BX(edt):
-        return int.from_bytes(edt, 'big')
-
-    EPC_FUNCTIONS = {
-    	0xA0: _30A0,
-        0xA1: _30A1,
-        0xA3: _30A3,
-        0xA4: _30A4,
-        0xA5: _30A5,
-        0xAA: _30AA,
-        0xB0: _30B0,
-        0xB3: _30BX,
-        0xBB: _30BX,
-        0xBE: _30BX,
-    }
-
-
     """
     GetOperationaTemperature get the temperature that has been set in the HVAC
